api/requirements.txt
fastapi
uvicorn[standard]
Pillow
numpy
scikit-learn
python-multipart
torch
